Moved all headers not directly needed into an "internal" folder.
