+++ include/Dice/hash/internal/wyhash.h
